+++ src/params.cxx
@@ -26,7 +26,7 @@
 #include "params.h"
 #include "exceptions.h"
 
-int parse_frequency(std::string s) {
+int64_t parse_frequency(std::string s) {
   std::istringstream ss(s);
   double f;
   std::string multiplier;
@@ -39,7 +39,7 @@ int parse_frequency(std::string s) {
     f *= 1e9;
   else if (multiplier != "")
     return -1;
-  return (int)f;
+  return (int64_t)f;
 }
 
 double parse_time(std::string s) {
